Move component to shared folder & Use it on BTC
